+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+- added button component
 - added support for having arrow on card content and on the top of the header
 - added card component (#15)
 - made styling to work in a more small (mobile) first kinda of way